code editor api cleanup circle now render as it should
